// src/app/utils/types/CustomerResponceTypes.ts
// Generated by https://quicktype.io

export interface CustomerResponceTypes {
	status: string;
	code: number;
	data: CustomerData[];
}

export interface CustomerData {
	status: Status;
	_id: string;
	customerName: string;
	customerContact: string;
	streetAddress: string;
	stateId: string;
	ABN: string;
	contacts: Contact[];
	createdAt: Date;
	updatedAt: Date;
	__v: number;
	postCode?: string;
	suburb?: null | string;
}

export interface Contact {
	_id: string;
	position: string;
	email: string;
	phone: string;
	title?: string;
}

export enum Status {
	Active = 'active'
}